requirements.txt
psutil~=7.1.2
requests~=2.32.5
feedparser~=6.0.12
PyYAML~=6.0.3
python-dotenv~=1.2.1
